+++ python/lsst/ts/ofc/__init__.py
@@ -25,7 +25,6 @@ try:
 except ModuleNotFoundError:
     pass
 
-from .camrot import *
 from .bend_mode_to_force import *
 from .sensitivity_matrix import *
 from .state_estimator import *