+++ tests/test_utils.py
@@ -23,7 +23,6 @@ import unittest
 import numpy as np
 
 from lsst.ts.ofc.utils import get_pkg_root, get_config_dir, rot_1d_array
-from lsst.ts.ofc import CamRot
 
 
 class TestUtils(unittest.TestCase):
@@ -38,31 +37,6 @@ class TestUtils(unittest.TestCase):
         config_dir = get_config_dir()
 
         self.assertTrue(config_dir.exists())
-
-    def test_rot_1d_array(self):
-        vec = np.array([1.0, 0.0])
-
-        angle = 0.0
-        rot_mat = CamRot.rot_mat(angle)
-        rot_vec = rot_1d_array(vec, rot_mat)
-
-        self.assertEqual(vec[0], rot_vec[0])
-        self.assertEqual(vec[1], rot_vec[1])
-
-        angle = 45.0
-        rot_mat = CamRot.rot_mat(angle)
-        rot_vec = rot_1d_array(vec, rot_mat)
-
-        expected_value = np.sin(np.radians(angle))
-        self.assertAlmostEqual(rot_vec[0], expected_value)
-        self.assertAlmostEqual(rot_vec[1], expected_value)
-
-        angle = 90.0
-        rot_mat = CamRot.rot_mat(angle)
-        rot_vec = rot_1d_array(vec, rot_mat)
-
-        self.assertAlmostEqual(rot_vec[0], vec[1])
-        self.assertAlmostEqual(rot_vec[1], vec[0])
 
 
 if __name__ == "__main__":